WIP: moving helper functions to lib/surface.py
untested
